+++ index.html
@@ -252,6 +252,8 @@
   </template>
 
   <script src="./js/main.js"></script>
+  <skript type="module" scr="main.js">
+    << /skript>
 </body>
 
 </html>
