+++ js/constants.js
@@ -1,0 +1,19 @@
+export const MIN_LIKES = 15;
+export const MAX_LIKES = 200;
+
+export const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо',
+  'Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.',
+];
+
+
+export const NAMES = [
+  'Егор',
+  'Иван',
+  'Василий',
+  'Александр',
+  'Дарья',
+  'Мария'
+];
